Spelling mistakes on Axminster blog post
Axminster Tools & Machinery is spelled as in Axminster in Devon, the town in which they are based.

Now I've read this more carefully this reads a bit like a clumsy advert placement on our blog, and I am sort-of against that. The person changes from "our members went" to "we enjoyed" before and after the photos.

Perhaps we could get some of the members who went to write about their experiences rather than just placing some token "it was nice" text in followed by a list of courses. We should also be upfront about the prices.